fixed dtype of inputs

+++ opencood/tools/inference_utils.py
@@ -63,7 +63,7 @@ def inference_early_fusion(batch_data, model, dataset):
     processed_lidar = cav_content['processed_lidar']
 
     voxel_features = processed_lidar['voxel_features'].to(torch.float32)
-    voxel_coords = processed_lidar['voxel_coords'].to(torch.float32)
+    voxel_coords = processed_lidar['voxel_coords'].to(torch.int32)
     voxel_num_points = processed_lidar['voxel_num_points'].to(torch.float32)
     record_len = cav_content['record_len'].to(torch.int32)
     spatial_correction_matrix = cav_content['spatial_correction_matrix'].to(torch.float32)
